p1-pcam-deployment/serving/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
boto3>=1.28.0
